requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
attrs==19.3.0
Django==3.0.2
django-graphql-jwt==0.3.0
Faker==0.9.1
graphene==2.1.8
graphene-django==2.8.0
graphql-core==2.2.1
graphql-relay==2.0.1
importlib-metadata==1.4.0
mixer==6.1.3
more-itertools==8.1.0
packaging==20.0
pluggy==0.13.1
promise==2.3
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.3
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.2
wcwidth==0.1.8
zipp==1.0.0
